add and modify 磁吸玩法更新： 1、移除了对铁矿方块的磁吸。 2、添加了磁吸玩法的配置文件设置。 3、增加对一些生物的磁吸。 4、现在能对携带了“一些用铁合成物品和能被磁吸的方块的物品”的生物进行磁吸了。

+++ src/main/java/com/ddhuan/ifscience/common/Entity/MagnetAttractedBlockEntity.java
@@ -25,7 +25,7 @@ public class MagnetAttractedBlockEntity extends BlockEntity {
     private Vector3d magnetAttractor_lastPos;//磁吸者上一个位置
     public double G = 0.04D;//重力加速度
     public int noAttractedTick = 0;//不被磁吸的时刻
-    public byte collisionFlag = magnetUtil.collisionFlag;
+    public int collisionFlag = magnetUtil.collisionFlag;
 
     public MagnetAttractedBlockEntity(EntityType<?> entityTypeIn, World worldIn) {
         super(entityTypeIn, worldIn);
